packages/yourloops/lib/custom-hooks/sort-comparators.hook.ts
/*
 * Copyright (c) 2023, Diabeloop
 *
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, are permitted provided that the following conditions are met:
 *
 * 1. Redistributions of source code must retain the above copyright notice, this
 *    list of conditions and the following disclaimer.
 *
 * 2. Redistributions in binary form must reproduce the above copyright notice,
 *    this list of conditions and the following disclaimer in the documentation
 *    and/or other materials provided with the distribution.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
 * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
 * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
 * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
 * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
 * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */

import { useUserName } from './user-name.hook'
import { type PatientProfile } from '../patient/models/patient-profile.model'
import { type GridComparatorFn } from '@mui/x-data-grid'

interface SortComparatorsHookReturn {
  sortByUserName: GridComparatorFn<PatientProfile>
}

export const useSortComparatorsHook = (): SortComparatorsHookReturn => {
  const { getUserName } = useUserName()

  const sortByUserName: GridComparatorFn<PatientProfile> = (patient1, patient2): number => {
    const patient1FullName = getUserName(patient1.firstName, patient1.lastName, patient1.fullName)
    const patient2FullName = getUserName(patient2.firstName, patient2.lastName, patient2.fullName)
    return patient1FullName.localeCompare(patient2FullName)
  }

  return { sortByUserName }
}
